refactor: move "bus" domain data structures to authly-domain
