+++ lib/authly-domain/Cargo.toml
@@ -32,6 +32,7 @@ reqwest = { version = "0.12", default-features = false, features = [
     "json",
     "rustls-tls",
 ] }
+serde = { version = "1", features = ["derive"] }
 rustls = { version = "0.23", default-features = false }
 thiserror = "2"
 time = "0.3"
