+++ src/directory.rs
@@ -3,8 +3,9 @@ use std::{any::Any, collections::HashMap, fmt::Display, fs};
 use authly_common::id::ServiceId;
 use authly_db::{Db, DbError};
 use authly_domain::{
+    bus::{BusError, ClusterMessage},
     cert::{client_cert, CertificateParamsExt},
-    ctx::{GetDb, GetDecryptedDeks, GetInstance},
+    ctx::{ClusterBus, GetDb, GetDecryptedDeks, GetInstance},
     directory::{DirKey, OAuthDirectory, PersonaDirectory},
     encryption::DecryptedDeks,
     id::BuiltinProp,
@@ -15,8 +16,6 @@ use tracing::error;
 
 use crate::{
     audit::Actor,
-    bus::{message::ClusterMessage, BusError},
-    ctx::ClusterBus,
     db::{
         cryptography_db::{self, CrDbError},
         directory_db::DbDirectory,
