+++ src/lib.rs
@@ -11,12 +11,15 @@ use std::{
 use anyhow::anyhow;
 use arc_swap::ArcSwap;
 use authly_domain::{
-    builtins::Builtins, ctx::GetDb, directory::PersonaDirectory, encryption::DecryptedDeks,
+    builtins::Builtins,
+    bus::ServiceMessage,
+    ctx::{GetDb, ServiceBus},
+    directory::PersonaDirectory,
+    encryption::DecryptedDeks,
     instance::AuthlyInstance,
 };
 use axum::{response::IntoResponse, Json};
-use bus::{message::ServiceMessage, service_events::ServiceEventDispatcher};
-use ctx::ServiceBus;
+use bus::service_events::ServiceEventDispatcher;
 use db::{cryptography_db, init_db, settings_db};
 use document::load::load_cfg_documents;
 pub use env_config::EnvConfig;
